rename application -> com.jiro4989.tkfm
