+++ build.gradle
@@ -30,5 +30,5 @@ dependencies {
 
 application {
     // Define the main class for the application.
-    mainClassName = 'com.jiro4989.tkfm.App'
+    mainClassName = 'com.jiro4989.tkfm.Main'
 }
