+++ src/main/java/com/jiro4989/tkfm/fileList/FileListHBox.java
@@ -1,8 +1,8 @@
-package application.fileList;
+package com.jiro4989.tkfm.fileList;
 
 import java.io.IOException;
 
-import application.MainController;
+import com.jiro4989.tkfm.MainController;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.layout.HBox;
