+++ src/main/java/com/jiro4989/tkfm/options/Numberings.java
@@ -1,4 +1,4 @@
-package application.options;
+package com.jiro4989.tkfm.options;
 
 import java.util.stream.IntStream;
