+++ src/main/java/com/jiro4989/tkfm/outputViewer/OutputViewerAnchorPane.java
@@ -1,8 +1,8 @@
-package application.outputViewer;
+package com.jiro4989.tkfm.outputViewer;
 
 import java.io.IOException;
 
-import application.MainController;
+import com.jiro4989.tkfm.MainController;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.layout.AnchorPane;
